+++ src/app/core/ui/navigation/menu-item/menu-item.component.scss
@@ -1,5 +1,5 @@
-@use "variables/sizes";
-@use "variables/colors";
+@use "../../../../../styles/variables/sizes";
+@use "../../../../../styles/variables/colors";
 
 /* ensures that all icons have the same width */
 .nav-icon {
